CacheMan can now take many backends
The backends each provide their own expiration time
and are looked through in their original order.

The same data is pushed into every backend implementation in order
The backend implementations are iterated through until the first one
found which declares valid data - that data is returned.

Timeout is removed from the options object and is now part of the
StorageBackend interface. Breaking change